requirements.txt
flask
flask_sqlalchemy
pandas
pandas_ta
yfinance
apscheduler
python-dotenv
python-telegram-bot